pyproject.toml
[tool.poetry]
name = "post-comments"
version = "0.1.0"
description = ""
authors = ["Alexandr Zibinin <zibinin.chel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.2"
black = "^24.10.0"
django = "^5.1.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
